Updates per review.
