+++ avail/src/main/kotlin/avail/anvil/tasks/AbstractWorkbenchModuleTask.kt
@@ -1,0 +1,59 @@
+/*
+ * AbstractWorkbenchModuleTask.kt
+ * Copyright © 1993-2023, The Avail Foundation, LLC.
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ *  * Redistributions of source code must retain the above copyright notice,
+ *    this list of conditions and the following disclaimer.
+ *
+ *  * Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ *
+ *  * Neither the name of the copyright holder nor the names of the contributors
+ *    may be used to endorse or promote products derived from this software
+ *    without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+ * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+ * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+ * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+ * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package avail.anvil.tasks
+
+import avail.anvil.AvailWorkbench
+import avail.builder.ResolvedModuleName
+import avail.descriptor.module.ModuleDescriptor
+
+/**
+ * An abstract [AbstractWorkbenchTask] involving a target
+ * [ResolvedModuleName].
+ *
+ * @property workbench
+ *   The owning [AvailWorkbench].
+ * @property targetModuleName
+ *   The resolved name of the target [module][ModuleDescriptor].
+ *
+ * @constructor
+ * Construct a new `AbstractWorkbenchModuleTask`.
+ *
+ * @param workbench
+ *   The owning [AvailWorkbench].
+ * @param targetModuleName
+ *   The resolved name of the target [module][ModuleDescriptor].
+ */
+abstract class AbstractWorkbenchModuleTask constructor(
+	workbench: AvailWorkbench,
+	protected val targetModuleName: ResolvedModuleName
+) : AbstractWorkbenchTask(workbench)
